Refactor repository structure: Moved database interaction functions from models to repository layer for better separation of concerns. Deleted old model files for budget, category, currency, transaction, and user. Added new repository files for each entity with corresponding CRUD operations.
